+++ src/components/BaseLineItem.vue
