+++ src/router/index.js
@@ -7,17 +7,17 @@ const routes = [
   {
     path: '/shop/test-tenant-slug',
     name: 'CustomerHome',
-    component: () => import('@/views/customers/Home.vue')
+    component: () => import('@/views/CustomerHome.vue')
   },
   {
     path: '/shop/test-tenant-slug/business-hours',
-    name: 'BusinessHours',
-    component: () => import('@/views/customers/BusinessHours.vue')
+    name: 'CustomerBusinessHours',
+    component: () => import('@/views/CustomerBusinessHours.vue')
   },
   {
     path: '/shop/test-tenant-slug/contact',
-    name: 'BusinessHours',
-    component: () => import('@/views/customers/Contact.vue')
+    name: 'CustomerContact',
+    component: () => import('@/views/CustomerContact.vue')
   },
   {
     path: '/',
