+++ src/views/CustomerHome.vue
@@ -1,6 +1,6 @@
 <template>
   <div id="customer-home" class="min-h-screen flex flex-col">
-    <MenuHeader />
+    <BaseMenuHeader />
     <div
       class="p-6 pb-20 flex-grow flex flex-col items-center justify-end text-on-background-image"
     >
@@ -19,15 +19,15 @@
 </template>
 
 <script>
-import MenuHeader from '@/components/MenuHeader.vue';
-import Button from '@/components/ui/Button.vue';
+import BaseMenuHeader from '@/components/BaseMenuHeader.vue';
+import Button from '@/components/BaseButton.vue';
 import CalendarIcon from '@/assets/icons/calendar_today.svg';
 import { mapGetters } from 'vuex';
 
 export default {
   name: 'CustomerHome',
   components: {
-    MenuHeader,
+    BaseMenuHeader,
     Button,
     CalendarIcon
   },
